refactor(html): tree construction "before head"
